Moved bgbkup-cli files to a new dir and ns.
