+++ .gitignore
@@ -11,7 +11,7 @@ build/
 vendor/
 package-lock.json
 yarn-error.log
-cron/bgbkup-cli.log
+cli/bgbkup-cli.log
 cron/cron-test.config
 cron/cron-test.result
 cron/restore-info.json
